milady: split benchmark and registry client modules
