+++ src/services/registry-client-network.ts
@@ -1,0 +1,148 @@
+import type { RegistryPluginInfo } from "./registry-client.js";
+
+export async function fetchFromNetwork(params: {
+  generatedRegistryUrl: string;
+  indexRegistryUrl: string;
+  applyLocalWorkspaceApps: (
+    plugins: Map<string, RegistryPluginInfo>,
+  ) => Promise<void>;
+  applyNodeModulePlugins: (
+    plugins: Map<string, RegistryPluginInfo>,
+  ) => Promise<void>;
+  sanitizeSandbox: (value?: string) => string;
+}): Promise<Map<string, RegistryPluginInfo>> {
+  const {
+    generatedRegistryUrl,
+    indexRegistryUrl,
+    applyLocalWorkspaceApps,
+    applyNodeModulePlugins,
+    sanitizeSandbox,
+  } = params;
+
+  try {
+    const resp = await fetch(generatedRegistryUrl, { redirect: "error" });
+    if (resp.ok) {
+      const data = (await resp.json()) as {
+        registry: Record<
+          string,
+          {
+            git: {
+              repo: string;
+              v0: { branch: string | null };
+              v1: { branch: string | null };
+              v2: { branch: string | null };
+            };
+            npm: {
+              repo: string;
+              v0: string | null;
+              v1: string | null;
+              v2: string | null;
+            };
+            supports: { v0: boolean; v1: boolean; v2: boolean };
+            description: string;
+            homepage: string | null;
+            topics: string[];
+            stargazers_count: number;
+            language: string;
+            kind?: string;
+            app?: {
+              displayName: string;
+              category: string;
+              launchType: string;
+              launchUrl: string | null;
+              icon: string | null;
+              capabilities: string[];
+              minPlayers: number | null;
+              maxPlayers: number | null;
+              viewer?: {
+                url: string;
+                embedParams?: Record<string, string>;
+                postMessageAuth?: boolean;
+                sandbox?: string;
+              };
+            };
+          }
+        >;
+      };
+      const plugins = new Map<string, RegistryPluginInfo>();
+      for (const [name, e] of Object.entries(data.registry)) {
+        const info: RegistryPluginInfo = {
+          name,
+          gitRepo: e.git.repo,
+          gitUrl: `https://github.com/${e.git.repo}.git`,
+          description: e.description || "",
+          homepage: e.homepage,
+          topics: e.topics || [],
+          stars: e.stargazers_count || 0,
+          language: e.language || "TypeScript",
+          npm: {
+            package: e.npm.repo,
+            v0Version: e.npm.v0,
+            v1Version: e.npm.v1,
+            v2Version: e.npm.v2,
+          },
+          git: {
+            v0Branch: e.git.v0?.branch ?? null,
+            v1Branch: e.git.v1?.branch ?? null,
+            v2Branch: e.git.v2?.branch ?? null,
+          },
+          supports: e.supports,
+        };
+
+        if (e.kind === "app" || e.app) {
+          info.kind = "app";
+        }
+        if (e.app) {
+          info.appMeta = {
+            displayName: e.app.displayName,
+            category: e.app.category,
+            launchType: e.app.launchType,
+            launchUrl: e.app.launchUrl,
+            icon: e.app.icon,
+            capabilities: e.app.capabilities || [],
+            minPlayers: e.app.minPlayers ?? null,
+            maxPlayers: e.app.maxPlayers ?? null,
+            viewer: e.app.viewer
+              ? {
+                  ...e.app.viewer,
+                  sandbox: sanitizeSandbox(e.app.viewer.sandbox),
+                }
+              : undefined,
+          };
+        }
+
+        plugins.set(name, info);
+      }
+      await applyLocalWorkspaceApps(plugins);
+      await applyNodeModulePlugins(plugins);
+      return plugins;
+    }
+  } catch {
+    // caller logs fallback warnings
+  }
+
+  const resp = await fetch(indexRegistryUrl, { redirect: "error" });
+  if (!resp.ok)
+    throw new Error(`index.json: ${resp.status} ${resp.statusText}`);
+  const data = (await resp.json()) as Record<string, string>;
+  const plugins = new Map<string, RegistryPluginInfo>();
+  for (const [name, gitRef] of Object.entries(data)) {
+    const repo = gitRef.replace(/^github:/, "");
+    plugins.set(name, {
+      name,
+      gitRepo: repo,
+      gitUrl: `https://github.com/${repo}.git`,
+      description: "",
+      homepage: null,
+      topics: [],
+      stars: 0,
+      language: "TypeScript",
+      npm: { package: name, v0Version: null, v1Version: null, v2Version: null },
+      git: { v0Branch: null, v1Branch: null, v2Branch: "next" },
+      supports: { v0: false, v1: false, v2: false },
+    });
+  }
+  await applyLocalWorkspaceApps(plugins);
+  await applyNodeModulePlugins(plugins);
+  return plugins;
+}
